extract shared code between query and thunk
